+++ run_test.sh
@@ -149,6 +149,15 @@ MYPY_PROTOBUF_VENV=venv_$PY_VER_MYPY_PROTOBUF
         echo -e "${RED}Some .pyi files did not match. Please commit those files${NC}"
         exit 1
     fi
+
+    # Generate 3rd party protos
+    mkdir -p third_party/out/generated_googleapis
+    # Known conflict with extensions proto in googleapis - skip that one
+    find third_party/googleapis -name "*.proto" \
+        ! -path "third_party/googleapis/google/cloud/compute/v1/compute.proto" \
+        ! -path "third_party/googleapis/google/cloud/compute/v1beta/compute.proto" \
+        ! -path "third_party/googleapis/google/cloud/compute/v1small/compute_small.proto" \
+        -print0 | xargs -0 "$PROTOC" --proto_path=third_party/googleapis --mypy_out=third_party/out/generated_googleapis --mypy_grpc_out=third_party/out/generated_googleapis --python_out=third_party/out/generated_googleapis
 )
 
 ERROR_FILE=$(mktemp)
@@ -180,6 +189,14 @@ for PY_VER in $PY_VER_UNIT_TESTS; do
         # Run async_only mypy
         ASYNC_ONLY_MODULES=( -m test.async_only.test_async_only )
         MYPYPATH=$MYPYPATH:test/generated_async_only mypy ${CUSTOM_TYPESHED_DIR_ARG:+"$CUSTOM_TYPESHED_DIR_ARG"} --report-deprecated-as-note --python-executable="$UNIT_TESTS_VENV"/bin/python --python-version="$PY_VER_MYPY_TARGET" "${ASYNC_ONLY_MODULES[@]}"
+
+        # Run google/protobuf mypy
+        GOOGLE_PROTOBUF=( test/generated/google/protobuf )
+        MYPYPATH=$MYPYPATH:test/generated PYTHONPATH=test/generated mypy --explicit-package-bases ${CUSTOM_TYPESHED_DIR_ARG:+"$CUSTOM_TYPESHED_DIR_ARG"} --report-deprecated-as-note --python-executable="$UNIT_TESTS_VENV"/bin/python --python-version="$PY_VER_MYPY_TARGET" "${GOOGLE_PROTOBUF[@]}"
+
+        # Run googleapis mypy
+        GOOGLEAPIS=( third_party/out/generated_googleapis )
+        MYPYPATH=$MYPYPATH:third_party/out/generated_googleapis mypy --explicit-package-bases ${CUSTOM_TYPESHED_DIR_ARG:+"$CUSTOM_TYPESHED_DIR_ARG"} --report-deprecated-as-note --python-executable="$UNIT_TESTS_VENV"/bin/python --python-version="$PY_VER_MYPY_TARGET" "${GOOGLEAPIS[@]}"
 
         export MYPYPATH=$MYPYPATH:test/generated
 
@@ -232,6 +249,10 @@ for PY_VER in $PY_VER_UNIT_TESTS; do
         PYTHONPATH=test/generated py.test --ignore=test/generated --ignore=test/generated_sync_only --ignore=test/generated_async_only --ignore=third_party -v
     )
 done
+
+
+# Clean up googleapis
+rm -rf third_party/out/generated_googleapis
 
 # Report all errors at the end
 if [ -s "$ERROR_FILE" ]; then