+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+- Use `builtins.property` to handle conflicts with fields named `property`
+- Mangle all non provided message type imports, this prevents conflicts with field names like `collections`, `builtins`, etc.
+  - Do not mangle message imports, as that would be a breaking change.
+
 ## 4.0.0
 
 - BREAKING: Drop support for `py_generic_services` as it was removed from the protobuf compiler starting in version 6.30